requirements.txt
Flask-Login==0.6.3
Jinja2==3.1.3
MarkupSafe==2.1.5
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
flask==3.0.2
importlib-metadata==7.0.1
itsdangerous==2.1.2
pymongo==4.6.2
python-dotenv==1.0.1
werkzeug==3.0.1
zipp==3.17.0